Refactored password provider to password handler:
The library does not provide any passwords, it only handles them. This
refactor is to enforce the explicitness.

Thanks to @Brunty for bringing this to my attention.
